application-backend-server/requirements.txt
flask
requests
python-jose[cryptography]
prometheus_client
mysql-connector-python
